Move FontAwesome online example
