+++ HtmlForgeX.Examples/Program.cs
@@ -10,6 +10,7 @@ using HtmlForgeX.Examples.Support;
 using HtmlForgeX.Examples.Forms;
 using HtmlForgeX.Examples.Tabler;
 using HtmlForgeX.Examples.VisNetwork;
+using HtmlForgeX.Examples.Icons;
 
 namespace HtmlForgeX.Examples;
 
@@ -36,6 +37,7 @@ internal class Program {
         // Tabler examples (console output)
         ExampleTablerIcon.Create();
         ExampleSvgIcons.Create();
+        ExampleFontAwesomeOnline.Create(openInBrowser);
 
         // Tabler examples html output
         ExampleTablerTag.Create(openInBrowser);
